requirements.txt
python-telegram-bot
httpcore
requests